rename client_state_download to client_reset_operation
